image.bbclass: make KERNEL_IMAGE & INITRD_IMAGE variable fixed
If the root file system does not contain the kernel or initrd file the
corresponding variables will be empty. This is the case when the root
file system is not yet generated or those files are deleted from it at
some point.

This patches sets those variables to a fixed value that is valid in
every state of the root file system.

Signed-off-by: Claudius Heine <ch@denx.de>