image-postproc-extension: add removal of machine-id
The machine-id(5) should contain a unique id of the machine, but since
this file is created by the systemd postinst script, every machine using
the same image would have the same id.

The file is removed in a postprocessing step and will be regenerated by
systemd on the first boot.

Signed-off-by: Claudius Heine <ch@denx.de>